+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 2026-09-02
+
+### Added
+
+- Added a config option for the default processing time of recipes when not explicitly defined.
+- Added support for custom `processing_time` definitions on individual manual mixing and grinding recipes.
+
+### Fixed
+
+- Millstones and Mortars now properly respect Create recipe fields like `processing_time`.
+
 ## [2.3.5] - 2026-08-31
 
 ### Fixed